pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["terabyte. <terabyte@terabyteis.me>"]
description = ""
name = "ModMail"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8.6"
jishaku = "^2.1.0"
repltable = "^0.1.4"
replit = "^3.2.4"
Flask = "^2.0.2"
discord-py-interactions = "^4.0.0"
humanize = "^4.0.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
mypy = "^0.910"
black = "^21.7b0"